+++ base_decorator.rb
