+++ nameable.rb
